# api/docs/swagger.yaml
basePath: /api
definitions:
  internal_urls.UrlShort:
    properties:
      long_url:
        type: string
    required:
    - long_url
    type: object
host: localhost:8080
info:
  contact: {}
  description: API to shorten long URLs into short codes.
  title: URL Shortner API
  version: "1.0"
paths:
  /{shortUrl}:
    get:
      description: Looks up the long URL for a short code and redirects the browser.
      parameters:
      - description: Short code to resolve
        in: path
        name: shortUrl
        required: true
        type: string
      responses:
        "302":
          description: Long URL to redirect to
          schema:
            type: header
        "404":
          description: Short code not found
          schema:
            additionalProperties: true
            type: object
      summary: Redirect a short code
      tags:
      - urls
  /url_shorter:
    post:
      consumes:
      - application/json
      description: Accepts a long URL and returns a short code. Retries on collisions.
      parameters:
      - description: Long URL to shorten
        in: body
        name: payload
        required: true
        schema:
          $ref: '#/definitions/internal_urls.UrlShort'
      produces:
      - application/json
      responses:
        "201":
          description: Shortened URL created
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad request
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Internal server error
          schema:
            additionalProperties: true
            type: object
      summary: Shorten a long URL
      tags:
      - urls
  /url_shorter/{shortUrl}:
    get:
      description: Looks up the long URL for a short code.
      parameters:
      - description: Short code to resolve
        in: path
        name: shortUrl
        required: true
        type: string
      produces:
      - application/json
      responses:
        "302":
          description: Redirect found
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Short code not found
          schema:
            additionalProperties: true
            type: object
      summary: Resolve a short code
      tags:
      - urls
swagger: "2.0"